// frontend/src/lib/api/generated/models/SearchEventsItem.ts
export class SearchEventsItem {
    'id': number;
    'slug': string;
    'organizationName': string;
    'name': string;
    'startAt': Date;
    'endAt': Date;
    'totalTicketNumber': number;
    'applicationStartAt': Date;
    'applicationEndAt': Date;
    'coverImageUrl': string;
    'organizeCityCode'?: string | null;
    'organizeAddress'?: string | null;
    'isOnline'?: boolean | null;
    'isOffline'?: boolean | null;
    'isBookmarked'?: boolean | null;
    'meetingToolCode'?: string | null;
    'soldTicketsNumber'?: number | null;
    'tags'?: Array<TagItem>;
    'publishedAt': Date;
    'minTicketPrice'?: number | null;
    'bookmarkCount'?: number | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": ""
        },
        {
            "name": "slug",
            "baseName": "slug",
            "type": "string",
            "format": ""
        },
        {
            "name": "organizationName",
            "baseName": "organization_name",
            "type": "string",
            "format": ""
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "startAt",
            "baseName": "start_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "endAt",
            "baseName": "end_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "totalTicketNumber",
            "baseName": "total_ticket_number",
            "type": "number",
            "format": ""
        },
        {
            "name": "applicationStartAt",
            "baseName": "application_start_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "applicationEndAt",
            "baseName": "application_end_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "coverImageUrl",
            "baseName": "cover_image_url",
            "type": "string",
            "format": ""
        },
        {
            "name": "organizeCityCode",
            "baseName": "organize_city_code",
            "type": "string",
            "format": ""
        },
        {
            "name": "organizeAddress",
            "baseName": "organize_address",
            "type": "string",
            "format": ""
        },
        {
            "name": "isOnline",
            "baseName": "is_online",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "isOffline",
            "baseName": "is_offline",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "isBookmarked",
            "baseName": "is_bookmarked",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "meetingToolCode",
            "baseName": "meeting_tool_code",
            "type": "string",
            "format": ""
        },
        {
            "name": "soldTicketsNumber",
            "baseName": "sold_tickets_number",
            "type": "number",
            "format": ""
        },
        {
            "name": "tags",
            "baseName": "tags",
            "type": "Array<TagItem>",
            "format": ""
        },
        {
            "name": "publishedAt",
            "baseName": "published_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "minTicketPrice",
            "baseName": "min_ticket_price",
            "type": "number",
            "format": ""
        },
        {
            "name": "bookmarkCount",
            "baseName": "bookmark_count",
            "type": "number",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return SearchEventsItem.attributeTypeMap;
    }

    public constructor() {
    }
}